Extracted MetaAnalysisReport component
